Move Dynamo helper to avoid circular import
